pyproject.toml
[project]
name = "conscience-servitor"
version = "0.1.0"
description = "EFHF Layer 5+ meta-orchestrator — pre-response ethical consistency enforcement via LLM2Vec-Gen"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastmcp>=3.1",
    "torch",
    "numpy>=1.26",
    "peft",
    "transformers",
    "huggingface-hub",
    "pyyaml",
]

[project.scripts]
conscience-servitor = "conscience_servitor.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/conscience_servitor"]
